+++ src/domain/membership/information/MembershipInformation.tsx
@@ -5,9 +5,13 @@ import { Button } from 'hds-react';
 
 import { MembershipInformation as MembershipInformationTypes } from '../../../graphql/generatedTypes';
 import LinkButton from '../../../common/components/linkButton/LinkButton';
-import getFullName from '../helpers/getFullName';
 import convertDateToLocale from '../../../common/helpers/convertDateToLocale';
+import getFullName from '../helpers/getFullName';
 import styles from './membershipInformation.module.css';
+
+const feedbackFormLink =
+  // eslint-disable-next-line max-len
+  'https://docs.google.com/forms/d/e/1FAIpQLSdqw2Lq3qooEeRdgr0sV0-Wv-4XcV7IZVzq1HuWoLRa2M7tEg/viewform?usp=pp_url&entry.1982410290=Nuorten+j%C3%A4senkortti';
 
 interface Props {
   onRenewMembership: () => void;
@@ -40,6 +44,7 @@ function MembershipInformation({
           </p>
           <QRCode
             size={175}
+            // eslint-disable-next-line max-len
             value={`${process.env.REACT_APP_ADMIN_URL}youthProfiles/${membershipInformationTypes.youthProfile?.profile.id}/show`}
           />
            
@@ -58,6 +63,13 @@ function MembershipInformation({
             path="/membership-details"
             component="Link"
             buttonText={t('membershipInformation.showProfileInformation')}
+            variant="secondary"
+          />
+          <LinkButton
+            className={styles.button}
+            path={feedbackFormLink}
+            component="a"
+            buttonText={t('feedback.giveFeedback.label')}
             variant="secondary"
           />
         </React.Fragment>