+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
+### Added
+- Temporary feedback link button in membership information page
+
 ### Changed
 - Use GitHub actions instead of Travis